finish add transaction to database, but temporary push to transactionBE

+++ app/src/main/java/com/example/mainproject/viewModel/AppViewModel.kt
@@ -96,7 +96,7 @@ class AppViewModel(
                         val transaction = transactionChildSnapshot.getValue(Transaction::class.java)
                         transaction?.let {
                             if (!it.isPositive) {
-                                totalExpense += it.amount
+                                totalExpense += it.amount.toDouble()
                             }
                             transactionListForDate.add(it)
                         }
